requirements.txt
alembic           1.15.2   
annotated-types   0.7.0    
anyio             4.9.0    
certifi           2025.1.31
click             8.1.8    
colorama          0.4.6    
dnspython         2.7.0    
email_validator   2.2.0    
fastapi           0.115.12
fastapi-cli       0.0.7
greenlet          3.1.1
h11               0.14.0
httpcore          1.0.7
httptools         0.6.4
httpx             0.28.1
idna              3.10
Jinja2            3.1.6
Mako              1.3.9
markdown-it-py    3.0.0
MarkupSafe        3.0.2
mdurl             0.1.2
pip               24.2
psycopg2          2.9.10
psycopg2-binary   2.9.10
pydantic          2.11.1
pydantic_core     2.33.0
Pygments          2.19.1
python-dotenv     1.1.0
python-multipart  0.0.20
PyYAML            6.0.2
rich              14.0.0
rich-toolkit      0.14.1
shellingham       1.5.4
sniffio           1.3.1
SQLAlchemy        2.0.40
starlette         0.46.1
typer             0.15.2
typing_extensions 4.13.0
typing-inspection 0.4.0
uvicorn           0.34.0
watchfiles        1.0.4
websockets        15.0.1